bootscript edits

+++ entrypoint.sh
@@ -1,10 +1,13 @@
 #!/bin/bash
 
-rm /run/dbus/pid &
-/etc/init.d/bluetooth start &
+if [[ -e "/run/dbus/pid" ]]; then
+	rm /run/dbus/pid &
+fi
+
+bluetooth start &
 /usr/bin/dbus-daemon --system &
 /usr/lib/bluetooth/bluetoothd &
-
+echo "bluetooth service started, moving on."
 
 if [[ -e "/config/setupdone" ]]; then
 	./scripts/autopair &
@@ -12,4 +15,5 @@ else
     ./scripts/setup &
 fi
 
+echo "All done starting snapclient.."
 snapclient -h $HOST
